lambda_src/requirements.txt
telethon==1.38.1           # For Telegram API interaction
python-dotenv==1.0.1       # For loading .env configurations
langchain-openai==0.3.0    # For using LangChain's OpenAI integration
openai==1.59.6             # For accessing OpenAI API
pytz==2024.2               # Time Zone conversion
requests==2.32.3           # Handle image downloads